requirements.txt
python-dotenv
motor
pyrogram
cachetools
requests
TgCrypto
